documentation + cleanup
