+++ specta-typescript/src/typescript.rs
@@ -2,7 +2,7 @@ use std::{borrow::Cow, io, path::Path};
 
 use specta::{datatype::DeprecatedType, TypeCollection};
 
-use crate::{comments, ExportError};
+use crate::{comments, primitives, ExportError};
 
 #[derive(Debug)]
 #[non_exhaustive]
@@ -135,17 +135,15 @@ impl Typescript {
         out += &self.framework_header;
         out.push_str("\n\n");
 
+        // TODO: Bring this back
         // if let Some((ty_name, l0, l1)) = detect_duplicate_type_names(&types).into_iter().next() {
         //     return Err(ExportError::DuplicateTypeName(ty_name, l0, l1));
         // }
 
-        // for (_, ty) in types.into_iter() {
-        //     is_valid_ty(&ty.inner, &types)?;
-
-        //     out += &export_named_datatype(self, ty, &types)?;
-        //     out += "\n\n";
-        // }
-        todo!();
+        for (_, ndt) in types.into_iter() {
+            out += &primitives::export(self, &types, ndt)?;
+            out += "\n\n";
+        }
 
         Ok(out)
     }
