+++ tests/tests/serde/internally_tagged.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use specta::Type;
-use specta_serde::SerdeError;
+use specta_serde::Error;
 
 use crate::ts::assert_ts;
